♻️ デバッグ機能をモーダル方式に変更
- debug-screenからBootstrap modalに変更
- DebugSceneクラス削除、BattleScene内に統合
- Game.tsからDebugState関連削除
- バトル状態を保持したままデバッグ可能
- 現在値が自動設定される直感的なUI実現
- プレイヤー詳細モーダルと同様のUX統一
